feature/decoration fix for review
