--- packages/client/src/entities/game/utils.ts
@@ -1,5 +1,0 @@
-import { CELL_SIZE } from './constants'
-
-export function toPixels(size: number): number {
-  return size * CELL_SIZE
-}
